// SDKs/twilio.com/twilio_trusthub_v1/1.37.3/typescript/src/sdk/models/shared/trusthubv1customerprofilecustomerprofileentityassignment.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class TrusthubV1CustomerProfileCustomerProfileEntityAssignment extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=customer_profile_sid" })
  customerProfileSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=object_sid" })
  objectSid?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}